Integrate mutlizone, display (#40) (#41)
* Log multizone and side tofs together

* Add screen

* Improve start procedure for display

* Show left/right tofs on logger: incorrect

* Cleanup

* Partial re-write of side tof library

* Show start screen for 2 seconds

* Driving animation

* Cleanup

* Cleanup

* Cleanup display

* Cleanup unused methods

* Variable standardization

---------

+++ src/networking/send_data_to_server.cpp
@@ -21,7 +21,7 @@ void SendDataToServer::attachIRData(JsonObject& payload) {
     JsonArray irArray = payload.createNestedArray("irSensorData");
     
     // Add each sensor reading to the array
-    for (int i = 0; i < 5; i++) {
+    for (uint8_t i = 0; i < 5; i++) {
         irArray.add(irSensorData[i]);
     }
 }
